requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
dateparser==1.2.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
numpy==1.26.2
pandas==2.1.4
pandas_ta==0.3.14b0
pycryptodome==3.19.1
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
six==1.16.0
tzdata==2023.4
tzlocal==5.2
ujson==5.9.0
urllib3==2.1.0
websockets==12.0
yarl==1.9.4
